+++ examples/console-utils-test.cpp
@@ -4,8 +4,8 @@
 // see for licence details (MIT) at its official distribution:
 // https://github.com/FOLF-projects/folf-namespace
 // ===========================================================
-#include <folf/timeTool.hpp>
-#include <folf/conTool.hpp>
+#include <folf/timeTools.hpp>
+#include <folf/conTools.hpp>
 #include <thread>
 
 void timer(bool *finished)
