+++ examples/prime-sum-benchmark.cpp
@@ -4,10 +4,10 @@
 // see for licence details (MIT) at its official distribution:
 // https://github.com/FOLF-projects/folf-namespace
 // ===========================================================
-#include <folf/timeTool.hpp>
+#include <folf/timeTools.hpp>
 #include <folf/info.hpp>
-#include <folf/numTool.hpp>
-#include <folf/conTool.hpp>
+#include <folf/calcTools.hpp>
+#include <folf/conTools.hpp>
 #include <iostream>
 #include <thread>
