+++ folf/conTools.hpp
